refactor settings service for sep concerns
