+++ back-end/config/db-job-config.js
@@ -2,6 +2,7 @@ require('./config');     //instantiate configuration variables
 require('./../global_functions'); //global functions for model defs
 require("./system_permissions"); //instantiate global permissions lists
 require('./model_constants'); //instantiate model constants
+require('./workflow_constants'); //instantiate workflow constants
 
 //DATABASE
 const models = require("./../models");
@@ -12,8 +13,8 @@ let dbPromise = models.sequelize.authenticate().then(() => {
     console.log('Performing startup migration...');
     return migratorPerform();
 }).then((migrations) => {
-    let workflowConstants = require('./workflow_constants');
-    workflowConstants.refreshSettingValues();
+    let settingsService = require('../services/SettingService');
+    settingsService.refreshSettingValues();
     return migrations;
 }).catch(err => {
         console.error('Unable prepare RDBMS for app:', process.env.DATABASE_URL, err);
