+++ back-end/services/SettingService.js
@@ -1,7 +1,5 @@
 'use strict';
-
 const Setting = require('../models').Setting;
-const workflowConstants = require('../config/workflow_constants');
 
 /** Changes value of specified setting.
  * 
@@ -17,7 +15,7 @@ const changeSettingValue = async function (setting_id, setting_value) {
   setting.value = setting_value;
   [err, setting] = await to(setting.save());
 
-  workflowConstants.refreshSettingValues();
+  refreshSettingValues();
 
   return setting;
 }
@@ -31,3 +29,52 @@ const getAllSettings = async function () {
   return settings;
 };
 module.exports.getAllSettings = getAllSettings;
+
+/** Is used to parse values depending on what their data type is defined by constant
+ * in SETTING_DATA_TYPES variable
+ */
+const parseValue = function (value, type) {
+  let parsedValue;
+  switch (type) {
+      case SETTING_DATA_TYPES.Integer:
+          parsedValue = parseInt(value, 10);
+          break;
+      case SETTING_DATA_TYPES.Float: 
+          parsedValue = parseFloat(value);
+          break;
+      case SETTING_DATA_TYPES.String:
+          parsedValue = value;
+          break;
+      case SETTING_DATA_TYPES.Boolean:
+          if (String(a) == "true")
+              parsedValue = true;
+          else
+              parsedValue = false;
+          break;
+      default: 
+          parsedValue = value;
+  }
+
+  return parsedValue;
+};
+
+/** Takes values from DB table Setting and defines them in global SYSTEM_SETTING variable.
+*  This function should be called after editing values of system settings.
+*/
+const refreshSettingValues = async () => {
+  let settings = await getAllSettings();
+  //try parse settings from DB, use default in case of error
+  if (settings.length)
+      settings.map(setting => {
+          try {
+              SYSTEM_SETTINGS[setting.key] = this.parseValue(setting.value, setting.type);
+          } catch(err) {
+              console.error(`Error parsing DB setting at ${setting.key}: ${err}. Using default value ${DEFAULT_SETTINGS[setting.key]}...`)
+              SYSTEM_SETTINGS[settings.key] = DEFAULT_SETTINGS[setting.key];
+          }
+
+      });
+  else
+      TE("Couldn't get settings values");
+};
+module.exports.refreshSettingValues = refreshSettingValues;
